refactor(template parser tests): move template humanizer utils into separate file, prepare for deferred test. Minor analyzer warning fixes.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=152159699
