+++ lib/src/common/forms/directives/ng_form.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
-import "package:angular2/core.dart"
-    show Directive, Provider, Optional, Inject, Self;
+import "package:angular2/core.dart" show Directive, Provider;
+import "package:angular2/di.dart" show Optional, Inject, Self;
 import "package:angular2/src/facade/async.dart" show EventEmitter;
 
 import "../model.dart" show AbstractControl, ControlGroup, Control;
