+++ lib/src/common/forms/directives/ng_form_model.dart
@@ -1,5 +1,6 @@
 import 'package:angular2/core.dart'
-    show SimpleChange, OnChanges, Directive, Provider, Inject, Optional, Self;
+    show SimpleChange, OnChanges, Directive, Provider;
+import "package:angular2/di.dart" show Optional, Inject, Self;
 import 'package:angular2/src/facade/async.dart' show EventEmitter;
 import 'package:angular2/src/facade/exceptions.dart' show BaseException;
