+++ test/compiler/template_parser_test.dart
@@ -15,11 +15,10 @@ import "package:logging/logging.dart";
 import 'package:test/test.dart';
 
 import "../test_util.dart";
-import "expression_parser/unparser.dart" show Unparser;
 import "schema_registry_mock.dart" show MockSchemaRegistry;
 import "test_bindings.dart" show TEST_PROVIDERS;
+import "template_humanizer_util.dart";
 
-var expressionUnparser = new Unparser();
 var someModuleUrl = "package:someModule";
 var MOCK_SCHEMA_REGISTRY = [
   provide(ElementSchemaRegistry,
@@ -28,6 +27,46 @@ var MOCK_SCHEMA_REGISTRY = [
 ];
 
 typedef R Func3Opt1<A, B, C, R>(A a, B b, [C c]);
+
+class ArrayConsole {
+  List<String> logs = [];
+  List<String> warnings = [];
+
+  ArrayConsole() {
+    Logger.root.onRecord.listen((LogRecord rec) {
+      if (rec.level == Level.WARNING) {
+        warn(rec.message);
+      } else {
+        log(rec.message);
+      }
+    });
+  }
+
+  void log(String msg) {
+    this.logs.add(msg);
+  }
+
+  void warn(String msg) {
+    this.warnings.add(msg);
+  }
+
+  void clear() {
+    logs.clear();
+    warnings.clear();
+  }
+}
+
+bool compareProviderList(List a, List b) {
+  if (a.length != b.length) return false;
+  for (int i = 0; i < a.length; i++) {
+    if (a[i]?.toJson()?.toString() != b[i]?.toJson()?.toString()) {
+      print('a> ${a[i]?.toJson()}');
+      print('b> ${b[i]?.toJson()}');
+      return false;
+    }
+  }
+  return true;
+}
 
 void main() {
   CompileDirectiveMetadata ngIf;
@@ -55,6 +94,7 @@ void main() {
       };
     });
   };
+
   group("TemplateParser", () {
     setUp(() async {
       beforeEachProviders(() => [
@@ -1737,303 +1777,4 @@ void main() {
       });
     });
   });
-}
-
-List<dynamic> humanizeTplAst(List<TemplateAst> templateAsts) {
-  var humanizer = new TemplateHumanizer(false);
-  templateVisitAll(humanizer, templateAsts);
-  return humanizer.result;
-}
-
-List<dynamic> humanizeTplAstSourceSpans(List<TemplateAst> templateAsts) {
-  var humanizer = new TemplateHumanizer(true);
-  templateVisitAll(humanizer, templateAsts);
-  return humanizer.result;
-}
-
-class TemplateHumanizer implements TemplateAstVisitor {
-  bool includeSourceSpan;
-  List<dynamic> result = [];
-  TemplateHumanizer(this.includeSourceSpan);
-  dynamic visitNgContent(NgContentAst ast, dynamic context) {
-    var res = [NgContentAst];
-    this.result.add(this._appendContext(ast, res));
-    return null;
-  }
-
-  dynamic visitEmbeddedTemplate(EmbeddedTemplateAst ast, dynamic context) {
-    var res = [EmbeddedTemplateAst];
-    this.result.add(this._appendContext(ast, res));
-    templateVisitAll(this, ast.attrs);
-    templateVisitAll(this, ast.outputs);
-    templateVisitAll(this, ast.references);
-    templateVisitAll(this, ast.variables);
-    templateVisitAll(this, ast.directives);
-    templateVisitAll(this, ast.children);
-    return null;
-  }
-
-  dynamic visitElement(ElementAst ast, dynamic context) {
-    var res = [ElementAst, ast.name];
-    this.result.add(this._appendContext(ast, res));
-    templateVisitAll(this, ast.attrs);
-    templateVisitAll(this, ast.inputs);
-    templateVisitAll(this, ast.outputs);
-    templateVisitAll(this, ast.references);
-    templateVisitAll(this, ast.directives);
-    templateVisitAll(this, ast.children);
-    return null;
-  }
-
-  dynamic visitReference(ReferenceAst ast, dynamic context) {
-    var res = [ReferenceAst, ast.name, ast.value];
-    this.result.add(this._appendContext(ast, res));
-    return null;
-  }
-
-  dynamic visitVariable(VariableAst ast, dynamic context) {
-    var res = [VariableAst, ast.name, ast.value];
-    this.result.add(this._appendContext(ast, res));
-    return null;
-  }
-
-  dynamic visitEvent(BoundEventAst ast, dynamic context) {
-    var res = [
-      BoundEventAst,
-      ast.name,
-      null, // TODO: remove
-      expressionUnparser.unparse(ast.handler)
-    ];
-    this.result.add(this._appendContext(ast, res));
-    return null;
-  }
-
-  dynamic visitElementProperty(BoundElementPropertyAst ast, dynamic context) {
-    var res = [
-      BoundElementPropertyAst,
-      ast.type,
-      ast.name,
-      expressionUnparser.unparse(ast.value),
-      ast.unit
-    ];
-    this.result.add(this._appendContext(ast, res));
-    return null;
-  }
-
-  dynamic visitAttr(AttrAst ast, dynamic context) {
-    var res = [AttrAst, ast.name, ast.value];
-    this.result.add(this._appendContext(ast, res));
-    return null;
-  }
-
-  dynamic visitBoundText(BoundTextAst ast, dynamic context) {
-    var res = [BoundTextAst, expressionUnparser.unparse(ast.value)];
-    this.result.add(this._appendContext(ast, res));
-    return null;
-  }
-
-  dynamic visitText(TextAst ast, dynamic context) {
-    var res = [TextAst, ast.value];
-    this.result.add(this._appendContext(ast, res));
-    return null;
-  }
-
-  dynamic visitDirective(DirectiveAst ast, dynamic context) {
-    var res = [DirectiveAst, ast.directive];
-    this.result.add(this._appendContext(ast, res));
-    templateVisitAll(this, ast.inputs);
-    templateVisitAll(this, ast.hostProperties);
-    templateVisitAll(this, ast.hostEvents);
-    return null;
-  }
-
-  dynamic visitDirectiveProperty(
-      BoundDirectivePropertyAst ast, dynamic context) {
-    var res = [
-      BoundDirectivePropertyAst,
-      ast.directiveName,
-      expressionUnparser.unparse(ast.value)
-    ];
-    this.result.add(this._appendContext(ast, res));
-    return null;
-  }
-
-  List<dynamic> _appendContext(TemplateAst ast, List<dynamic> input) {
-    if (!this.includeSourceSpan) return input;
-    input.add(ast.sourceSpan.text);
-    return input;
-  }
-}
-
-String sourceInfo(TemplateAst ast) {
-  return '${ast.sourceSpan.text}: ${ast.sourceSpan.start.offset}';
-}
-
-List<dynamic> humanizeContentProjection(List<TemplateAst> templateAsts) {
-  var humanizer = new TemplateContentProjectionHumanizer();
-  templateVisitAll(humanizer, templateAsts);
-  return humanizer.result;
-}
-
-class TemplateContentProjectionHumanizer implements TemplateAstVisitor {
-  List<dynamic> result = [];
-  dynamic visitNgContent(NgContentAst ast, dynamic context) {
-    this.result.add(["ng-content", ast.ngContentIndex]);
-    return null;
-  }
-
-  dynamic visitEmbeddedTemplate(EmbeddedTemplateAst ast, dynamic context) {
-    this.result.add(["template", ast.ngContentIndex]);
-    templateVisitAll(this, ast.children);
-    return null;
-  }
-
-  dynamic visitElement(ElementAst ast, dynamic context) {
-    this.result.add([ast.name, ast.ngContentIndex]);
-    templateVisitAll(this, ast.children);
-    return null;
-  }
-
-  dynamic visitReference(ReferenceAst ast, dynamic context) {
-    return null;
-  }
-
-  dynamic visitVariable(VariableAst ast, dynamic context) {
-    return null;
-  }
-
-  dynamic visitEvent(BoundEventAst ast, dynamic context) {
-    return null;
-  }
-
-  dynamic visitElementProperty(BoundElementPropertyAst ast, dynamic context) {
-    return null;
-  }
-
-  dynamic visitAttr(AttrAst ast, dynamic context) {
-    return null;
-  }
-
-  dynamic visitBoundText(BoundTextAst ast, dynamic context) {
-    this.result.add([
-      '''#text(${ expressionUnparser . unparse ( ast . value )})''',
-      ast.ngContentIndex
-    ]);
-    return null;
-  }
-
-  dynamic visitText(TextAst ast, dynamic context) {
-    this.result.add(['''#text(${ ast . value})''', ast.ngContentIndex]);
-    return null;
-  }
-
-  dynamic visitDirective(DirectiveAst ast, dynamic context) {
-    return null;
-  }
-
-  dynamic visitDirectiveProperty(
-      BoundDirectivePropertyAst ast, dynamic context) {
-    return null;
-  }
-}
-
-class FooAstTransformer implements TemplateAstVisitor {
-  dynamic visitNgContent(NgContentAst ast, dynamic context) {
-    throw "not implemented";
-  }
-
-  dynamic visitEmbeddedTemplate(EmbeddedTemplateAst ast, dynamic context) {
-    throw "not implemented";
-  }
-
-  dynamic visitElement(ElementAst ast, dynamic context) {
-    if (ast.name != "div") return ast;
-    return new ElementAst("foo", [], [], [], [], [], [], null, [],
-        ast.ngContentIndex, ast.sourceSpan);
-  }
-
-  dynamic visitReference(ReferenceAst ast, dynamic context) {
-    throw "not implemented";
-  }
-
-  dynamic visitVariable(VariableAst ast, dynamic context) {
-    throw "not implemented";
-  }
-
-  dynamic visitEvent(BoundEventAst ast, dynamic context) {
-    throw "not implemented";
-  }
-
-  dynamic visitElementProperty(BoundElementPropertyAst ast, dynamic context) {
-    throw "not implemented";
-  }
-
-  dynamic visitAttr(AttrAst ast, dynamic context) {
-    throw "not implemented";
-  }
-
-  dynamic visitBoundText(BoundTextAst ast, dynamic context) {
-    throw "not implemented";
-  }
-
-  dynamic visitText(TextAst ast, dynamic context) {
-    throw "not implemented";
-  }
-
-  dynamic visitDirective(DirectiveAst ast, dynamic context) {
-    throw "not implemented";
-  }
-
-  dynamic visitDirectiveProperty(
-      BoundDirectivePropertyAst ast, dynamic context) {
-    throw "not implemented";
-  }
-}
-
-class BarAstTransformer extends FooAstTransformer {
-  dynamic visitElement(ElementAst ast, dynamic context) {
-    if (ast.name != "foo") return ast;
-    return new ElementAst("bar", [], [], [], [], [], [], null, [],
-        ast.ngContentIndex, ast.sourceSpan);
-  }
-}
-
-bool compareProviderList(List a, List b) {
-  if (a.length != b.length) return false;
-  for (int i = 0; i < a.length; i++) {
-    if (a[i]?.toJson()?.toString() != b[i]?.toJson()?.toString()) {
-      print('a> ${a[i]?.toJson()}');
-      print('b> ${b[i]?.toJson()}');
-      return false;
-    }
-  }
-  return true;
-}
-
-class ArrayConsole {
-  List<String> logs = [];
-  List<String> warnings = [];
-
-  ArrayConsole() {
-    Logger.root.onRecord.listen((LogRecord rec) {
-      if (rec.level == Level.WARNING) {
-        warn(rec.message);
-      } else {
-        log(rec.message);
-      }
-    });
-  }
-
-  void log(String msg) {
-    this.logs.add(msg);
-  }
-
-  void warn(String msg) {
-    this.warnings.add(msg);
-  }
-
-  void clear() {
-    logs.clear();
-    warnings.clear();
-  }
 }
