+++ test/core/linker/security_integration_test.dart
@@ -3,7 +3,8 @@ library angular2.test.core.linker.security_integration_test;
 
 import 'dart:html';
 
-import 'package:angular2/core.dart' show provide, Injectable, OpaqueToken;
+import 'package:angular2/core.dart' show provide, OpaqueToken;
+import "package:angular2/di.dart" show Injectable;
 import 'package:angular2/src/core/metadata.dart' show Component, View;
 import 'package:angular2/src/security/dom_sanitization_service.dart';
 import 'package:angular2/src/testing/internal.dart';
